🐛Fix: Proof에서 중복 확인으로 수정

See merge request s11-ai-image-sub1/S11P21A101!154

+++ backend/src/main/java/com/a101/ecofarming/proof/repository/ProofRepository.java
@@ -18,6 +18,6 @@ public interface ProofRepository extends JpaRepository<Proof, Integer> {
 
     List<Proof> findByChallengeAndUser(Challenge challenge, User user);
 
-    Long countByChallengeAndUserAndIsValidTrue(Challenge challengeId, User userId);
+    Integer countByChallengeAndUserAndIsValidTrue(Challenge challengeId, User userId);
 }
